+++ NEWS.md
@@ -1,3 +1,9 @@
+# weathercan (development version)
+
+# weathercan 0.7.2
+- Fix normals to work with new ECCC data format
+- Prepare `normals_dl()` and family for new 1991-2020 normals
+
 # weathercan 0.7.1
 - `stations()` now uses the most recent version of the data even if it hasn't changed
   (prevent message regarding age of stations data frame).